src/microservices/modalities-service/requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
aio-pika==9.4.0
psycopg2-binary==2.9.11
sqlalchemy==2.0.44
prometheus-fastapi-instrumentator==7.1.0
python-logging-loki==0.3.1
alembic==1.11.1
authlib==1.3.0
python-multipart==0.0.6
httpx==0.24.0
